+++ config/config.php
@@ -18,5 +18,7 @@ return [
     'column_names' => [
 
         'model_morph_key' => 'model_id',
-    ]
+    ],
+
+    'middleware_key' => 'role',
 ];