+++ database/factories/RoleFactory.php
@@ -1,22 +1,29 @@
 <?php
 
+namespace Assignee\Factories;
+
 use Assignee\Models\Role;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+class RoleFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Role::class;
 
-$factory->define(Role::class, function (Faker $faker) {
-    return [
-        'name' => $faker->words('1', true),
-        'guard_name' => 'web',
-    ];
-});
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->words('1', true),
+            'guard_name' => 'web',
+        ];
+    }
+}