+++ src/Contracts/Package.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Assignee\Contracts;
+
+interface Package
+{
+
+    public function version();
+}