+++ src/Package.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Assignee;
+
+class Package
+{
+    const VERSION = '1.1.1';
+
+    public function version()
+    {
+        return self::VERSION;
+    }
+}