+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-assignee` will be documented in this file
 
+##  1.1.1 - 2020-09-26
+
+- Add Package Model
+- Auto register 'role' middleware
+
 ##  1.1.0 - 2020-09-08
 
 - Laravel 8.0